пофиксил вроде

+++ src/main/java/arena/CrawlerVars.java
@@ -10,8 +10,8 @@ public class CrawlerVars {
 
     /** Wave when it's over. */
     public static final int bossWave = 25;
-    /** Interval between waves in seconds. */
-    public static final int firstWaveDelay = 15, waveDelay = 10, additionalDelay = 2;
+    /** Interval between waves in ticks. */
+    public static final float firstWaveDelay = 15f, waveDelay = 10f, additionalDelay = 2f;
 
     /** The maximum number of units that players can spawn. */
     public static final int unitCap = 48;
